+++ src/tmtccmd/com_if/ethernet_com_if.py
@@ -15,7 +15,7 @@ from tmtccmd.utility.tmtcc_logger import get_logger
 from tmtccmd.com_if.com_interface_base import CommunicationInterface, PusTmListT
 from tmtccmd.pus_tm.factory import PusTelemetryFactory
 from tmtccmd.pus_tc.base import PusTcInfoT
-from tmtccmd.utility.tmtcc_tmtc_printer import TmTcPrinter
+from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.core.definitions import ethernet_address_t
 
 LOGGER = get_logger()
