+++ src/tmtccmd/com_if/qemu_com_if.py
@@ -28,7 +28,7 @@ from collections import deque
 from threading import Thread
 from tmtccmd.com_if.com_interface_base import CommunicationInterface, PusTcInfoT, PusTmListT
 from tmtccmd.pus_tm.factory import PusTelemetryFactory
-from tmtccmd.utility.tmtcc_tmtc_printer import TmTcPrinter
+from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.com_if.serial_com_if import SerialComIF, SerialCommunicationType
 from tmtccmd.utility.tmtcc_logger import get_logger
 from tmtccmd.utility.dle_encoder import encode_dle, decode_dle, STX_CHAR, ETX_CHAR, DleErrorCodes
