+++ src/tmtccmd/sendreceive/cmd_sender_receiver.py
@@ -15,7 +15,7 @@ import time
 
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
 from tmtccmd.core.definitions import QueueCommands, CoreGlobalIds
-from tmtccmd.utility.tmtcc_tmtc_printer import TmTcPrinter
+from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.utility.tmtcc_logger import get_logger
 
 from tmtccmd.sendreceive.tm_listener import TmListener
