+++ src/tmtccmd/sendreceive/sequential_sender_receiver.py
@@ -10,7 +10,7 @@ import time
 from tmtccmd.sendreceive.cmd_sender_receiver import CommandSenderReceiver
 from tmtccmd.sendreceive.tm_listener import TmListener
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
-from tmtccmd.utility.tmtcc_tmtc_printer import TmTcPrinter
+from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.utility.tmtcc_logger import get_logger
 from tmtccmd.pus_tc.base import TcQueueT
