+++ src/tmtccmd/sendreceive/single_command_sender_receiver.py
@@ -13,7 +13,7 @@ from tmtccmd.sendreceive.tm_listener import TmListener
 
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
 
-from tmtccmd.utility.tmtcc_tmtc_printer import TmTcPrinter
+from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.utility.tmtcc_logger import get_logger
 
 from tmtccmd.pus_tc.base import PusTcTupleT
